pyproject.toml
[project]
name = "cryptopian_playground_hantastic"
version = "0.0.1"
authors = [
    { name = "Han", email = "xjohnwu@gmail.com" },
]
dependencies = [
    "pycryptodome",
    "jsonpath-ng",
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"